Solving the Solar Cloud error
